--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,7 +29,7 @@ cast = "0.2"
 num-traits = { version = "0.2", default-features = false }
 oorandom = "11.1"
 rayon = "1.3"
-regex = { version = "1.3", default-features = false, features = ["std"] }
+regex = { version = "1.3", default-features = false, features = ["std"], optional = true }
 
 [dependencies.plotters]
 version = "^0.2.12"
@@ -49,7 +49,9 @@ maintenance = { status = "passively-maintained" }
 
 [features]
 real_blackbox = []
-default = []
+filter_regex = ["regex"]
+default = ["filter_regex"]
+
 
 [workspace]