--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,8 +15,8 @@ exclude = ["book/*"]
 
 [dependencies]
 lazy_static = "1.4"
-criterion-plot = { path="plot", version="0.4.1" }
-itertools = "0.8"
+criterion-plot = { path="plot", version="0.4.1", optional = true }
+itertools = { version = "0.8", optional = true }
 serde = "1.0"
 serde_json = "1.0"
 serde_derive = "1.0"
@@ -24,9 +24,9 @@ atty = "0.2"
 clap = { version = "2.33", default-features = false }
 csv = "1.1"
 walkdir = "2.3"
-tinytemplate = "1.0"
+tinytemplate = { version = "1.0", optional = true }
 cast = "0.2"
-num-traits = { version = "0.2", default-features = false }
+num-traits = { version = "0.2", default-features = false, features = ["std"] }
 oorandom = "11.1"
 rayon = "1.3"
 regex = { version = "1.3", default-features = false, features = ["std"], optional = true }
@@ -34,7 +34,8 @@ regex = { version = "1.3", default-features = false, features = ["std"], optiona
 [dependencies.plotters]
 version = "^0.2.12"
 default-features = false
-features = ["svg", "area_series", "line_series"] 
+features = ["svg", "area_series", "line_series"]
+optional = true
 
 [dev-dependencies]
 tempdir = "0.3.7"
@@ -50,8 +51,8 @@ maintenance = { status = "passively-maintained" }
 [features]
 real_blackbox = []
 filter_regex = ["regex"]
-default = ["filter_regex"]
-
+plotting = ["criterion-plot", "itertools", "tinytemplate", "plotters"]
+default = ["filter_regex", "plotting"]
 
 [workspace]