--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,7 +22,7 @@ serde_json = "1.0"
 serde_derive = "1.0"
 atty = "0.2"
 clap = { version = "2.33", default-features = false }
-csv = "1.1"
+csv = { version = "1.1", optional = true }
 walkdir = "2.3"
 tinytemplate = { version = "1.0", optional = true }
 cast = "0.2"
@@ -52,7 +52,7 @@ maintenance = { status = "passively-maintained" }
 real_blackbox = []
 filter_regex = ["regex"]
 plotting = ["criterion-plot", "itertools", "tinytemplate", "plotters"]
-default = ["filter_regex", "plotting"]
+default = ["filter_regex", "plotting", "csv"]
 
 [workspace]