@bheisler 2020-04-25
Yes, the dependency tree is quite large. I've tried to keep it as small as I can, but I feel your pain.

The way I intend to deal with this is to create a separate `cargo criterion` subcommand that invokes the benchmark executables and generates the plots. Then the benchmark executables themselves don't need to link in nearly so much code, and could be limited to just command-line and machine-readable output (probably using serde-json). The benchmarks would be much more lightweight, and the user would only have to compile the plot-generation code when they install the subcommand.

Unfortunately, I don't think I can hide existing code behind a cargo feature (even a default one) without breaking semver compatibility. If I'm going to do that, I might as well break it only once to introduce the subcommand.