@thomcc 2020-04-27
> Unfortunately, I don't think I can hide existing code behind a cargo feature (even a default one) without breaking semver compatibility.

I'm pretty sure this is not true. Many crates have made changes like this, including big ones like `serde`. Consider changes like adding no_std support, and then putting `std` behind a default-on feature.