@bheisler 2020-06-29
Hey, thanks again for the suggestion. I have some plans in this direction myself in the next breaking-change release - see #426 - but for the 0.3.* series I'll leave things as they are.